About panel now dockable
